componentized all of about page. Also moved about page to about folder with other components needed for page
